deleted excess server files
